\docType{data}
\name{vie}
\alias{vie}
\title{Vienna's Wiener Linien Network Data}
\format{A class \code{ntw} object, with roughly 150 lines and 1630 stations.}
\source{
  Provided by Wiener Linien under a CC-BY license. Data
  source: Stadt Wien data.wien.gv.at
}
\description{
  A class \code{ntw} object with the network meta data of
  Vienna's Wiener Linien public transit network. For
  details see \code{\link{getNetwork.vie}}.
}
\details{
  This data is provided by the Vienna city government. It
  has been added to the package during development time,
  and is most likely outdated. Use
  \code{\link{getNetwork.vie}} to obtain a newer version.
  See also the
  \href{https://open.wien.at/site/datensatz/?id=add66f20-d033-4eee-b9a0-47019828e698}{original
  data set description}.
}
\keyword{dataset}
\keyword{datasets}

